Re-organized project - put all function prototypes in vis.h and all general purpose functions in vislib.cpp
