+++ policy.csv
@@ -1,10 +1,10 @@
 1,bea@gmail.com,Hyundai ix35,Grey,JER749,19-06-2022,19-06-2023,C,50000,500,5000,104,W
 2,bea@gmail.com,Mazda CX5,Soul Red,GAG501,14-06-2022,14-06-2023,F,30000,500,2400,100,F
 6,gag@gmail.com,Hyundai Tucson,Black,GET432,17-06-2022,17-06-2023,C,40000,500,4000,83,W
-8,gag@gmail.com,BMW X5,Red,GAG440,19-06-2022,19-06-2023,C,60000,500,6000,125,W
+8,gag@gmail.com,Tesla,Black,GET567,23-06-2022,23-06-2023,C,80000,500,8000,166,W
 11,gag@gmail.com,Tesla,Black,GET560,19-06-2022,19-06-2023,C,80000,500,8000,166,W
-12,gag@gmail.com,Toyota RAV4,Black,TEG560,19-06-2022,19-06-2023,c,60000,500,6000,125,w
 13,bea@gmail.com,Tesla,Black,GET890,19-06-2022,19-06-2023,C,50000,500,5000,104,W
 18,gag@gmail.com,Honda CRV,Red,UYT321,19-06-2022,19-06-2023,C,60000,500,6000,125,W
 19,gag@gmail.com,Tesla,Black,ERT890,19-06-2022,19-06-2023,C,80000,500,8000,333,F
 20,gil@gmail.com,Tesla,Red,GET901,19-06-2022,19-06-2023,C,70000,500,7000,145,W
+22,gag@gmail.com,Toyota CHR,Yellow,GET789,23-06-2022,23-06-2023,C,60000,500,6000,125,W
