+++ users.csv
@@ -1,6 +1,4 @@
-gag@gmail.com,klmn,Gilberto Adarne,Gabon,0275581168,4
-bea@gmail.com,klmn,Gabon,Bea Gabriele,90901,1
-gag@yahoo.com,klmn,Gilberto,Gabon,90901,1
-gil@gmail.com,klmn,Gabon,Gil,1234567,1
-bob@gmail.com,klmn,Vargas,Bobby,123456,1
-gil@yahoo.com,klmn,Gabon,Gil,8989,1
+gag@gmail.com,klmn;<=,Gabon,Gilberto Adarne,0275581168,4
+bea@gmail.com,klmn,Bea Gabriele,Gabon,90901,1
+gag@yahoo.com,klmn,Gabon,Gilberto,90901,1
+gil@gmail.com,klmn,Gil,Gabon,1234567,1
